requirements.txt
discord.py==2.0.1
pymongo==4.2.0
python-dotenv==0.20.0
